requirements.txt
openai>=1.0
pyyaml
pandas
tqdm
